feat: :sparkles: Now teachers is able to grade studens assignments

+++ lib/task.ts
@@ -3,12 +3,74 @@ import { z } from "zod";
 import prisma from "./db";
 import { revalidatePath } from "next/cache";
 
+export async function changeTaskAssignment(formData: FormData) {
+    "use server";
+
+    const taskAssignment = formData.get("taskAssignment") as string;
+    const grade = formData.get("grade") as string;
+    const teamId = parseInt(formData.get("teamId") as string);
+    const taskId = parseInt(formData.get("taskId") as string);
+
+    // If grade equals delete, delete task assignment
+    if (grade === "delete") {
+        await prisma.taskAssignment.delete({
+            where: {
+                id: parseInt(taskAssignment),
+            },
+        });
+
+        // Redirect to team page
+        revalidatePath("/team/" + teamId + "/task/" + taskId);
+        return;
+    }
+
+    // If grade equals pending, set grade to null
+    if (grade === "pending") {
+        await prisma.taskAssignment.update({
+            where: {
+                id: parseInt(taskAssignment),
+            },
+            data: {
+                grade: null,
+            },
+        });
+
+        // Redirect to team page
+        revalidatePath("/team/" + teamId + "/task/" + taskId);
+        return;
+    }
+
+    await prisma.taskAssignment.update({
+        where: {
+            id: parseInt(taskAssignment),
+        },
+        data: {
+            grade: grade,
+        },
+    });
+
+    // Redirect to team page
+    revalidatePath("/team/" + teamId + "/task/" + taskId);
+}
+
 export async function deleteTaskAssignment(formData: FormData) {
     "use server";
 
     const taskAssignment = formData.get("taskAssignment") as string;
     const teamId = parseInt(formData.get("teamId") as string);
     const taskId = parseInt(formData.get("taskId") as string);
+
+    // Return if task assignment allready has a grade
+    const taskAssignmentData = await prisma.taskAssignment.findUnique({
+        where: {
+            id: parseInt(taskAssignment),
+        },
+    });
+
+    if (taskAssignmentData?.grade) {
+        revalidatePath("/team/" + teamId + "/task/" + taskId);
+        return;
+    }
 
     // Delete task assignment
     await prisma.taskAssignment.delete({
